requirements.txt
django-user-accounts==1.3.1
pinax-webanalytics==2.0.4
pinax-eventlog==1.1.2
requests
Django==1.10
pinax-theme-bootstrap==7.10.1
